pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "my first telegram bot"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot>=22.1",
]
